update all dependencies

+++ src/render/nav.rs
@@ -6,7 +6,7 @@ pub fn render_nav(current_page: u32, count: usize, limit: usize, template: &str)
             nav {
                 @let min = 1;
                 @let max = count / limit + 1;
-                @let nav_start = std::cmp::max(min as i32, current_page as i32 - 3);
+                @let nav_start = std::cmp::max(min, current_page as i32 - 3);
                 @let nav_end = std::cmp::min(max as i32, nav_start + 7);
                 a href=(format!("{}{}", template, min)) { "<<" }
                 @for page in nav_start..=nav_end {
